Pass int # of sections rather than float

+++ kaleid/KaleidView.swift
@@ -1,19 +1,19 @@
 import SwiftUI
 
 struct KaleidView<Content: View>: View {
-  let count: CGFloat
+  let count: Int
   let content: Content
 
-  init(count: CGFloat, @ViewBuilder _ content: () -> Content) {
+  init(count: Int, @ViewBuilder _ content: () -> Content) {
     self.count = count
     self.content = content()
   }
 
   var body: some View {
-    let angle = 360.0 / count * 2.0
+    let angle = 360.0 / CGFloat(count) * 2.0
 
     return ZStack {
-      ForEach(0..<Int(count / 2.0), id: \.self) { index in
+      ForEach(0..<Int(CGFloat(count) / 2.0), id: \.self) { index in
         MirroredView {
           content
             .clipShape(Pie(count: Int(count)))
